* Implement WalletAppService

* Remove unnecessary abstractions

* Remove unnecessary interface

* Refactor AesWalletEncryption to use asynchronous serialization

* Convert to sats/vKB properly

* Fix registrations

+++ src/Angor/Avalonia/Angor.Wallet/Domain/BroadcastedTransaction.cs
@@ -9,7 +9,7 @@ public record BroadcastedTransaction(
     IEnumerable<TransactionAddressInfo> AllOutputs, 
     long Fee,
     bool IsConfirmed,
-    int? BlockHeight,
+    long? BlockHeight,
     DateTimeOffset? BlockTime,
     string RawJson
 );
